create useOpenRecordInCommandMenu hook
